add central object management
this is needed as xpdacq.py needs to be "fed" in right objects
we need to isolate global object function from xpdacq.py, otherwise we can't call these objects
