+++ xpdacq/object_manage.py
@@ -1,0 +1,38 @@
+##############################################################################
+#
+# xpdacq            by Billinge Group
+#                   Simon J. L. Billinge sb2896@columbia.edu
+#                   (c) 2016 trustees of Columbia University in the City of
+#                        New York.
+#                   All rights reserved
+#
+# File coded by:    Timothy Liu, Simon Billinge
+#
+# See AUTHORS.txt for a list of people who contributed.
+# See LICENSE.txt for license information.
+#
+#
+##############################################################################
+
+def _areaDET(area_det_name):
+    global AREA_DET
+    AREA_DET = area_det_name
+
+def _tempController(temp_controller_name):
+    global TEMP_CONTROLLER
+    TEMP_CONTROLLER = temp_controller_name
+
+def _shutter(shutter_name):
+    global SHUTTER
+    SHUTTER = shutter_name
+
+def _bdir(b_dir_name):
+    global B_DIR
+    B_DIR = b_dir_name
+
+# analysis objects : not ready yet
+# db
+# LiveTable
+# get_events
+# get_images
+
